move styling
