+++ packages/grafana-ui/src/components/index.ts
@@ -11,7 +11,7 @@ export { default as resetSelectStyles } from './Select/resetSelectStyles';
 
 // Forms
 export { FormLabel } from './FormLabel/FormLabel';
-export { FormField } from './FormGroup/FormField';
+export { FormField } from './FormField/FormField';
 export { Label } from './Label/Label';
 
 export { LoadingPlaceholder } from './LoadingPlaceholder/LoadingPlaceholder';
